% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer_plus.R
\name{beamer_plus}
\alias{beamer_plus}
\title{Beamer presentation with additional arguments}
\usage{
beamer_plus(..., themeoptions = "default", colorthemeoptions = "default",
  fontthemeoptions = "default", fontsize = "11pt",
  classoption = list("aspectratio=1610"))
}
\description{
Beamer presentation with additional arguments
}

